+++ tests/factories/ConsumerFactory.php
@@ -1,0 +1,17 @@
+<?php
+
+use Faker\Generator as Faker;
+use Illuminate\Support\Str;
+use Vanilo\Cart\Tests\Dummies\Consumer;
+
+$factory->define(
+    Consumer::class, function (Faker $faker) {
+    static $password;
+
+    return [
+        'name'           => $faker->name,
+        'email'          => $faker->unique()->safeEmail,
+        'password'       => $password ?: $password = bcrypt('secret'),
+        'remember_token' => Str::random(10),
+    ];
+});